pyproject.toml
[project]
name = "ticksonic"
version = "0.1.0"
description = ""
authors = [
    {name = "Yamir Encarnacion"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-socketio (>=5.5.1,<6.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "websocket-client (>=1.8.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "polygon-api-client (>=1.14.3,<2.0.0)",
    "termcolor (>=2.5.0,<3.0.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "numpy (>=2.2.2,<3.0.0)",
    "databento (>=0.48.0,<0.49.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"